// src/app/shared/model/cust-bank-stmnt-obj.model.ts
export class CustBankStmntObj {
    CustBankStmntId: any;
    CustBankAccId: any;
    BegBalanceAmt: any;
    RowVersion: any;
    Month: any;
    Year: any;
    DebitTrxCount: any;
    DebitAmt: any;
    CreditTrxCount: any;
    CreditAmt: any;
    BalanceAmt: any;
    constructor(){this.CustBankStmntId = 0, this.RowVersion = ""}
}